Implement Authorization and Authentication
Login

Generate Json Web Token

Validate token

Protect some routes to allow user to acces their own resources only

+++ cookery-service/src/main/java/service/resources/RecipesResources.java
@@ -21,26 +21,15 @@ public class RecipesResources {
 
 
 	@GET //GET at http://localhost:XXXX/recipes?ingredient=onion OR http://localhost:XXXX/recipes
-	@Produces(MediaType.APPLICATION_JSON)
 	@PermitAll
-	public Response getRecipeByIngredient(@DefaultValue("all") @QueryParam("ingredient") String ingredient, @HeaderParam("Authorization") String auth){
-//		System.out.println("AUTH");
-//		System.out.println(auth);
-
-//		String credentials2 = new String(Base64.getDecoder().decode(auth.getBytes()));
-//		System.out.println("DECODED");
-//		System.out.println(credentials2);
-
+	@Produces(MediaType.APPLICATION_JSON)
+	public Response getRecipeByIngredient(@DefaultValue("all") @QueryParam("ingredient") String ingredient){
 		List<Recipe> recipes;
 		Controller controller = new Controller();
 		if(ingredient.equals("all")) {
-//			recipes = dataStore.getRecipes();
-
 			recipes = controller.getRecipes();
 		}
 		else {
-//			recipes = dataStore.getRecipesBy(ingredient);
-
 			recipes = controller.getRecipes(ingredient);
 		}
 
@@ -50,11 +39,10 @@ public class RecipesResources {
 	}
 
 	@GET //GET at http://localhost:XXXX/recipes/1
-	@Produces(MediaType.APPLICATION_JSON)
 	@Path("{id}")
 	@PermitAll
+	@Produces(MediaType.APPLICATION_JSON)
 	public Response getRecipe(@PathParam("id") int id) {
-//		Recipe recipe = dataStore.getRecipe(id);
 		Controller controller = new Controller();
 
 		Recipe recipe = controller.getRecipe(id);
@@ -71,8 +59,6 @@ public class RecipesResources {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@RolesAllowed({"user", "admin"})
 	public Response addRecipe(Recipe recipe){
-
-//		boolean result = dataStore.addRecipe(recipe);
 		Controller controller = new Controller();
 		controller.createRecipe(recipe);
 
@@ -85,16 +71,19 @@ public class RecipesResources {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Path("{id}")
 	@RolesAllowed({"user", "admin"})
-	public Response updateRecipe(@PathParam("id") int id, Recipe recipe){
-//		boolean result = dataStore.updateRecipe(id, recipe);
-
+	public Response updateRecipe(@PathParam("id") int id, Recipe recipe, @HeaderParam("Authorization") String auth){
 		Controller controller = new Controller();
 
+		int userId = controller.getIdInToken(auth); // id in token
+		int ownerId = controller.getUserId(id); // id of owner of the recipe
 
-		boolean result = controller.updateRecipe(id, recipe);
+		if(userId != ownerId) {
+			return Response.status(Response.Status.FORBIDDEN).entity("You are not allowed to perform this action").build();
+		}
 
-		System.out.println("recipe received: "+ recipe);
-		System.out.println("REACHED");
+		boolean result;
+
+		result = controller.updateRecipe(id, recipe);
 
 		if(result) {
 			return Response.noContent().build();
@@ -107,9 +96,16 @@ public class RecipesResources {
 	@DELETE //DELETE at http://localhost:XXXX/recipes/3
 	@Path("{id}")
 	@RolesAllowed({"user", "admin"})
-	public Response deleteRecipe(@PathParam("id") int id){
-//		dataStore.deleteRecipe(id);
+	public Response deleteRecipe(@PathParam("id") int id, @HeaderParam("Authorization") String auth){
 		Controller controller = new Controller();
+
+		int userId = controller.getIdInToken(auth); // id in token
+		int ownerId = controller.getUserId(id); // id of owner of the recipe
+
+		if(userId != ownerId) {
+			return Response.status(Response.Status.FORBIDDEN).entity("You are not allowed to perform this action").build();
+		}
+
 		controller.deleteRecipe(id);
 
 		return Response.noContent().build();
